+++ demos/chinook/src/main/java/is/codion/framework/demos/chinook/model/TrackTableModel.java
@@ -81,14 +81,14 @@ public final class TrackTableModel extends SwingEntityTableModel {
 		}
 
 		@Override
-		public Optional<ConditionModel<?>> createColumnCondition(Attribute<?> attribute) {
+		public Optional<ConditionModel<?>> create(Attribute<?> attribute) {
 			if (attribute.equals(Track.MEDIATYPE_FK)) {
 				return Optional.of(SwingForeignKeyConditionModel.builder(Track.MEDIATYPE_FK)
 								.includeEqualOperators(createEqualComboBoxModel(Track.MEDIATYPE_FK))
 								.build());
 			}
 
-			return super.createColumnCondition(attribute);
+			return super.create(attribute);
 		}
 	}
 }